Added tests for patient. Try change logic for note

+++ src/main/java/com/cleverdev/clientService/service/PatientService.java
@@ -1,5 +1,6 @@
 package com.cleverdev.clientService.service;
 
+import com.cleverdev.clientService.exceptions.PatientNotFoundException;
 import com.cleverdev.clientService.service.converter.PatientConvert;
 import com.cleverdev.clientService.entity.Patient;
 import com.cleverdev.clientService.exceptions.GuidAlreadyExistException;
@@ -7,6 +8,7 @@ import com.cleverdev.clientService.model.PatientModel;
 import com.cleverdev.clientService.repository.PatientRepository;
 import lombok.RequiredArgsConstructor;
 import org.json.simple.JSONArray;
+import org.springframework.core.ParameterizedTypeReference;
 import org.springframework.stereotype.Service;
 
 
@@ -16,20 +18,19 @@ import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class PatientService implements GetJsonFromOldSystem {
+public class PatientService {
+    ParameterizedTypeReference<JSONArray> typeRef = new ParameterizedTypeReference<>() {
+    };
     private final PatientRepository patientRepository;
     private final PatientConvert patientConvert;
 
-    public void createPatient(PatientModel patientModel) throws GuidAlreadyExistException {
-        if (patientRepository.findByOldClientGuid(patientModel.getOldClientGuid()) == null) {
+    public boolean createPatient(PatientModel patientModel) throws GuidAlreadyExistException {
+        if ((patientRepository.findByOldClientGuid(patientModel.getOldClientGuid())) == null) {
             patientRepository.save(patientConvert.fromPatientModelToPatient(patientModel));
-        } else {
-            Patient patient = patientRepository.findByOldClientGuid(patientModel.getOldClientGuid());
-            if (patient.getOldClientGuid().equals(patientModel.getOldClientGuid())) {
-                throw new GuidAlreadyExistException("EXCEPTION : Patient with this guid exists, unique values for type guid are required");
-            }
+            return true;
+        } else
+            throw new GuidAlreadyExistException("EXCEPTION : Patient with this guid exists, unique values for type guid are required");
         }
-    }
 
     public Patient pathMappingPatient(PatientModel patientModel, String guid) {
         Patient patient = patientRepository.findByOldClientGuid(guid);
@@ -41,14 +42,18 @@ public class PatientService implements GetJsonFromOldSystem {
         return patient;
     }
 
-    public Patient deletePatientFromDb(Long id) {
+    public Patient getPatientFromDb(Long id) throws PatientNotFoundException {
+        if (patientRepository.getById(id) != null) {
+            return patientRepository.getById(id);
+        }
+        throw new PatientNotFoundException("This patient not found!");
+    }
+    public Patient deletePatientFromDb(Long id) throws PatientNotFoundException {
+        if (patientRepository.getById(id) == null)
+            throw new PatientNotFoundException("This patient not found!");
+
         Patient patient = patientRepository.getById(id);
         patientRepository.deleteById(id);
         return patient;
-    }
-
-    @Override
-    public JSONArray getJsonObjFromOldSystem(String urlClient) {
-        return null;
     }
 }
